requirements.txt
streamlit>=1.45.0
plotly>=6.0.0
requests>=2.32.0
pandas>=2.2.0
python-dotenv>=1.1.0
fpdf2>=2.8.0
kaleido>=0.4.1
